refactor: switched to factory convention from clients
