+++ rhoas/acls/resource_acl.go
@@ -97,9 +97,9 @@ func aclCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.
 	// Warning or errors can be collected in a slice type
 	var diags diag.Diagnostics
 
-	api, ok := m.(rhoasAPI.Clients)
+	api, ok := m.(rhoasAPI.Factory)
 	if !ok {
-		return diag.Errorf("unable to cast %v to rhoasAPI.Clients", m)
+		return diag.Errorf("unable to cast %v to rhoasAPI.Factory", m)
 	}
 
 	kafkaID, ok := d.Get(KafkaIDField).(string)
