+++ rhoas/tests/serviceaccount_test.go
@@ -118,7 +118,7 @@ func testAccCheckServiceAccountDestroy(s *terraform.State) error {
 	// retrieve the connection established in Provider configuration
 	api, ok := testAccRHOAS.Meta().(rhoasAPI.Clients)
 	if !ok {
-		return errors.Errorf("unable to cast %v to rhoasAPI.Clients)", testAccRHOAS.Meta())
+		return errors.Errorf("unable to cast %v to rhoasAPI.Factory)", testAccRHOAS.Meta())
 	}
 
 	// loop through the resources in state, verifying each widget of type rhoas_kafka is destroyed
@@ -157,7 +157,7 @@ func testAccCheckServiceAccountExists(resource string, serviceAccount *saclient.
 
 		api, ok := testAccRHOAS.Meta().(rhoasAPI.Clients)
 		if !ok {
-			return errors.Errorf("unable to cast %v to rhoasAPI.Clients)", testAccRHOAS.Meta())
+			return errors.Errorf("unable to cast %v to rhoasAPI.Factory)", testAccRHOAS.Meta())
 		}
 		gotServiceAccount, resp, err := api.ServiceAccountMgmt().GetServiceAccount(context.Background(), rs.Primary.ID).Execute()
 		if err != nil {
